# Optimized format with compile time format strings