@quickfur 2018-01-11
Very interesting.

Just out of curiosity, what heuristic did you use for the various guessed length values for each %-spec?