@JackStouffer 2018-01-11
For strings I just choose 12 because it seemed reasonable.

For numbers, basically what I did is saw how many chars the number 999 took and used that.