@JackStouffer 2018-01-16
Going ahead and merging this.