@wilzbach 2018-01-16
I don't know why CodeCov doesn't display its status updates here (I have contacted their support), but this diff isn't well covered:

![image](https://user-images.githubusercontent.com/4370550/34969863-3a07ca30-fa71-11e7-888d-9d637207031b.png)

I am sorry for being so late to the party, but it shouldn't be too hard to evaluate `guessLength` at runtime, s.t. it's properly annotated with code coverage.