@JackStouffer 2018-01-16
Going forward with this again. Any other issues can be resolved in follow ups.